requirements.txt
gpiozero
python-dotenv
daemonize